refactor: time 패키지 reservation 내부로 이동
