+++ src/main/java/roomescape/reservation/domain/Reservation.java
@@ -17,7 +17,7 @@ import roomescape.common.validate.Validator;
 import roomescape.reservation.exception.PastDateReservationException;
 import roomescape.reservation.exception.PastTimeReservationException;
 import roomescape.theme.domain.Theme;
-import roomescape.time.domain.ReservationTime;
+import roomescape.reservation.time.domain.ReservationTime;
 import roomescape.user.domain.UserId;
 
 import java.time.LocalDateTime;
