+++ src/main/java/roomescape/reservation/time/domain/ReservationTimeId.java
@@ -1,4 +1,4 @@
-package roomescape.time.domain;
+package roomescape.reservation.time.domain;
 
 import roomescape.common.domain.EntityId;
