+++ src/main/java/roomescape/reservation/time/ui/ReservationTimeController.java
@@ -1,4 +1,4 @@
-package roomescape.time.ui;
+package roomescape.reservation.time.ui;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -10,9 +10,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.common.uri.UriFactory;
-import roomescape.time.application.ReservationTimeFacade;
-import roomescape.time.ui.dto.CreateReservationTimeWebRequest;
-import roomescape.time.ui.dto.ReservationTimeResponse;
+import roomescape.reservation.time.application.ReservationTimeFacade;
+import roomescape.reservation.time.ui.dto.CreateReservationTimeWebRequest;
+import roomescape.reservation.time.ui.dto.ReservationTimeResponse;
 
 import java.net.URI;
 import java.util.List;
