+++ src/main/java/roomescape/reservation/time/ui/dto/CreateReservationTimeWebRequest.java
@@ -1,27 +1,21 @@
-package roomescape.time.application.dto;
+package roomescape.reservation.time.ui.dto;
 
 import lombok.AccessLevel;
 import lombok.experimental.FieldNameConstants;
 import roomescape.common.domain.DomainTerm;
 import roomescape.common.validate.Validator;
-import roomescape.time.domain.ReservationTime;
 
 import java.time.LocalTime;
 
 @FieldNameConstants(level = AccessLevel.PRIVATE)
-public record CreateReservationTimeServiceRequest(LocalTime startAt) {
+public record CreateReservationTimeWebRequest(LocalTime startAt) {
 
-    public CreateReservationTimeServiceRequest {
+    public CreateReservationTimeWebRequest {
         validate(startAt);
     }
 
-    public ReservationTime toDomain() {
-        return ReservationTime.withoutId(
-                startAt);
-    }
-
     private void validate(final LocalTime startAt) {
-        Validator.of(CreateReservationTimeServiceRequest.class)
+        Validator.of(CreateReservationTimeWebRequest.class)
                 .validateNotNull(Fields.startAt, startAt, DomainTerm.RESERVATION_TIME.label());
     }
 }
