+++ src/test/java/roomescape/reservation/domain/ReservationTest.java
@@ -12,8 +12,8 @@ import roomescape.theme.domain.ThemeDescription;
 import roomescape.theme.domain.ThemeId;
 import roomescape.theme.domain.ThemeName;
 import roomescape.theme.domain.ThemeThumbnail;
-import roomescape.time.domain.ReservationTime;
-import roomescape.time.domain.ReservationTimeId;
+import roomescape.reservation.time.domain.ReservationTime;
+import roomescape.reservation.time.domain.ReservationTimeId;
 import roomescape.user.domain.User;
 import roomescape.user.domain.UserId;
 import roomescape.user.domain.UserName;
