+++ src/test/java/roomescape/time/domain/ReservationTimeIdTest.java
@@ -3,6 +3,7 @@ package roomescape.time.domain;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import roomescape.common.validate.InvalidArgumentException;
+import roomescape.reservation.time.domain.ReservationTimeId;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
